Update App.js and DropDownForm.js
